Cleaned up repo

+++ lib/models/data_structures/ds_robot_connection_data.py
@@ -5,5 +5,5 @@ class DsRobotConnectionData:
     name: str = None
     ip_address: str = None
     port: int = None
-    read_freq: int = None
+    read_freq: float = None
     is_deleted: bool = False
